+++ packages/components/src/Toast/index.ts
@@ -10,3 +10,4 @@
  * graphic logo are (registered/a) trademark(s) of Plan International.
  */
 export * from './Toast'
+export * from './ErrorToastNotification'
