+++ packages/components/src/interface/index.ts
@@ -20,5 +20,4 @@ export * from './ViewData'
 export * from './EventTopBar'
 export * from './ListTable'
 export * from './LocationSearch/LocationSearch'
-export * from './ErrorToastNotification'
 export * from './TableView'
